+++ src/index.mjs
@@ -6,12 +6,14 @@ import terrain50_parse from './static/Terrain50Parse.mjs';
 import terrain50_parse_stream from './static/Terrain50ParseStream.mjs';
 import terrain50_merge from './static/Terrain50Merge.mjs';
 import terrain50_validate from './static/Terrain50Validate.mjs';
+import terrain50_analyse_frequencies from './static/Terrain50AnalyseFrequencies.mjs';
 
 Terrain50.Blank = terrain50_blank;
 Terrain50.Parse = terrain50_parse;
 Terrain50.ParseStream = terrain50_parse_stream;
 Terrain50.Merge = terrain50_merge;
 Terrain50.Validate = terrain50_validate;
+Terrain50.AnalyseFrequencies = terrain50_analyse_frequencies;
 
 export {
 	Terrain50,